Overhaul image resize dialog
Summary:
D12300 added two percentage spinboxes, making the dialog look somewhat
crowded. It was also unclear which of the fields affected the image's
height, and which the width.

To polish the dialog and make it easier and faster to use, the following
changes were implemented:
- Change window title to "Resize Image".
- Remove redundant intro text.
- Add width and height labels.
- Don't use an italic font face for the current size.
- Align everything to the center of the dialog.
- Add "px" suffix to all sizes.
- Use 99,999 instead of 100,000 for the maximum size to optimize the width of the spinboxes.
- Drop redundant "Percentage" label.
- Don't show unnecessary decimals in the percentage spinboxes.
- Change maximum value for percentage to a more even 1000%.
- If "Keep aspect ratio" is enabled, disable the then unnecessary right percentage spinbox.
- Use a single grid layout instead of multiple sub-layouts.
- Prevent resizing the dialog, as with the default size everything should fit just fine.
- Set proper tabstops.

Depends on D12300

Test Plan:
Resize dialog (press {key Shift R} to open) should work as before, and
the checkbox should affect the state of the height percentage spinbox.

Before:

{F5818159}

After:

{F5818160}

Reviewers: #gwenview, huoni

Reviewed By: #gwenview, huoni

Subscribers: muhlenpfordt, ngraham, huoni

Differential Revision: https://phabricator.kde.org/D12458