Copy raster image and SVG content to clipboard
Summary:
The normal `Copy`/`Cut` options ({key Ctrl C}/{key Ctrl X}) copy the
url(s) of the selected file(s)/folder(s) to the clipboard.
This patch additionally copies raster image and SVG content of the
current document to the clipboard. In Browse Mode the content is
copied only if a single file is selected.

FEATURE: 290294

Test Plan:
# Open raster image or SVG in Gwenview
# {nav Edit > Copy} or {key Ctrl C} to copy image to clipboard
# Paste image to different applications (e.g. gimp, office writer,
  facebook chat in web browser, inkscape, etc.)
# Cut&Paste should only move the file if pasted to file manager
  applications (e.g. dolphin)

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: huoni, rkflx, ngraham

Differential Revision: https://phabricator.kde.org/D12871