Update paste action on current directory and selection changes
Summary:
The paste action is disabled on Gwenview start. This is caused by
initializing the action state before setting the current directory
url in `ContextManager`. The state updates only on clipboard changes
and does not reflect the current permissions to paste into a folder.
This patch updates the paste action state on changing the directory
and the selection, which both affect the paste destination.

BUG: 276255
FIXED-IN: 18.04.2

Test Plan:
# Copy url(s) to clipboard (e.g. from `Dolphin`)
# Open Gwenview in Browse Mode
# Check if {nav Edit > Paste...} is enabled on startup
# Check if {nav Edit > Paste...} is disabled in non-writable folders

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx, huoni

Differential Revision: https://phabricator.kde.org/D12902