Refine behaviour for dragging modified images to external applications
Summary:
In f365a38fa28f support for copying modified images to external
applications was added. With the refactoring in D13248, we can extend
that support for dragging.

Depends on D13248

Test Plan:
Start dragging in Gwenview and drop onto the target for every
combination of:
- Source: {nav Browse} mode, {nav Thumbnail Bar} (normal and fullscreen), {nav Start > Recent Files}
- Formats: JPG, PNG, NEF, SVG
- Regular and {nav Cropped} image for every source file format
- Target: Dolphin, GIMP, LibreOffice, Inkscape
Only dragging of modified images to Dolphin is known to be broken
currently, but no regression either.

Reviewers: #gwenview, muhlenpfordt

Reviewed By: #gwenview, muhlenpfordt

Differential Revision: https://phabricator.kde.org/D13249