Watch for installation of kipi-plugins only when needed
Summary:
When `kipi-plugins` was not installed and the user clicked on the
{nav Plugins} menu at least once, a filesystem watcher was added
unconditionally. This resulted in a huge amount of signals and
unnecessary updates in Gwenview as soon as any software installation
took place. In particular this was bad for packages with lots of files
like the kernel headers.

It's enough to add the watcher only in case the user actively clicked on
{nav Install Plugins}.

FIXED-IN: 18.04.2

Test Plan:
Remove `kipi-plugins`, make sure {nav Discover} is installed, start
Gwenview and click on {nav Plugins > Install Plugins}. Check that the
list of plugins still updates itself as soon as the installation
completes successfully.

Reviewers: #gwenview, ngraham

Reviewed By: ngraham

Differential Revision: https://phabricator.kde.org/D13250