Fix share menu not showing the first time it is used
Summary:
When the share menu is clicked for the first time (kipi plugins are not loaded
yet) the share menu doesn't work because updateMenu() clears the menu before any
plugin is available, thus the menu is empty and nothing is shown.

updateMenu() should be called when init() is triggered by loadingFinished() and not
when it's triggered by aboutToShow().

Test Plan:
- Start an new instance of Gwenview
- Click on 'Share'
- The share menu should show the first time without having to click 'Share' a second
time

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Differential Revision: https://phabricator.kde.org/D13289