Prompt to install kipi-plugins when the share button is clicked
Summary:
When 'kipi-plugins' is not installed, the 'Share' menu shows 'No Plugin Found'. With this patch it will present a link to install the missing plugins instead (Similarly to what the 'Plugins' menu does).

BUG: 386753

Test Plan:
Without the **kipi-plugins** package installed, click on Share in the Gwenview toolbar.
It should display a menu action 'Install Plugins' which triggers the AppStream URL (appstream://photolayoutseditor.desktop).

Reviewers: #gwenview, ngraham, rkflx

Reviewed By: #gwenview, ngraham, rkflx

Subscribers: rkflx, nicolasfella

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D13197