Do not allow to drag an image onto itself
Summary:
The combination of 131d25855e11 and 984b9737f079 led to the user being
able to drag an image in {nav View} mode and immediately drop it again.
The drag cursor implied Gwenview allowed this kind of operation, while
actually nothing would happen as the URL did not change.

To prevent confusion, we do not allow to drop the image in that
situation anymore. It is still possible to drop multiple images which
include the current image, so a future patch could implement switching
to {nav Compare} mode in that case.

As a side effect a small visual bug is fixed: When moving very slowly
while initiating a drag, the `ForbiddenCursor` would appear briefly
until you moved the mouse further along and the actual drag cursor
appeared. With the patch applied, the cursor will always keep being set
to `ForbiddenCursor`, preventing any flickering.

CCBUG: 386034

Test Plan:
Slowly start to drag an image in {nav View} mode. The cursor should
indicate no dropping is allowed.

Reviewers: #gwenview, huoni

Reviewed By: #gwenview, huoni

Subscribers: huoni

Differential Revision: https://phabricator.kde.org/D13724