Fix integer underflow warning in ImageMetaInfoModel
`QModelIndex::internalId` returns `quintptr`, which is `unsigned`,
conflicting with the negative indexes assigned to
`GroupRow::NoGroupSpace` and `GroupRow::NoGroup`. The latter are used
for marking through `internalId` when a row refers to an entry (i.e.
nested) instead of a group (i.e. top-level).

Since `GroupRow` also functions as an index for `mMetaInfoGroupVector`,
in addition to removing the negative values both `NoGroupSpace` and
`GroupSpace` have to be moved to the end of `GroupRow`.

Test Plan:
Click on {nav More} in the {nav Information} sidebar. {nav General} and
{nav EXIF} should be populated as before. Compiler warning should be
gone.