Allow to double-click to accelerate crop and red-eye modifications
Summary:
Currently double-clicking will switch to fullscreen when a tool like
{nav Crop} or {nav Reduce Red Eye} is used. It would be handy to allow
to accept the modifications without having to tediously position the
cursor over the corresponding button each time.

By using double-click to accept just like it's done for Spectacle's
{nav Rectangular Region} capture mode, users can improve the efficiency of
their primarily mouse-based workflow. In particular fixing multiple red
eyes, e.g. in a group picture, will become much faster. In GIMP's
{nav Crop}, clicking also accepts (albeit a single click is enough).

Triggering fullscreen can still be done before opening the tool, with a
shortcut or with the toolbutton, and for {nav Crop} when clicking
outside the selection. Furthermore, there is the possibility to
{nav Undo} if the double-click was unintentional.

To make users aware of the feature, for {nav Reduce Red Eye} the inline
help text is changed (at the same time removing the to-be-avoided "you"
in the string). For {nav Crop} the docbook will be adapted, since there
is not enough space in the UI for a subtle inline pointer (in a later
patch).

Test Plan:
Double-click in the {nav Crop} and {nav Reduce Red Eye} tools after
modifying the image. Gwenview should apply the changes immediately.

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: muhlenpfordt, ngraham

Differential Revision: https://phabricator.kde.org/D14755