Fix Advanced settings crop toolbar sometimes cut off
Summary:
When the {nav Crop} tool is started in standard mode and then enabling
`Advanced settings` the window width is not updated, which can result
in a cut off toolbar.
This is caused by the design of the image tools which are not embedded
by a layout to the image view and therefore not automatically
exchanging layout/resize events.
To solve this `SlideContainer::eventFilter()` now handles
`LayoutRequest` events of the content widgets.
Updating the geometry for these events makes some layout problems in
`SaveBar` visible which is also derived from `SlideContainer`.
The message and link labels need to be adjusted to be still shrinkable
and don't force the SaveBar getting too wide.

FIXED-IN: 18.08.1

Test Plan:
* Open {nav Crop} without checked `Advanced settings` option
* Set `Advanced settings` and check if window width is adjusted to
  show the whole toolbar
* Modify 3 or more images to enable SaveBar and show all possible
  parts (buttons, info and links)

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx

Differential Revision: https://phabricator.kde.org/D14917