Center numbers in Crop and Reduce Red Eye spinboxes
Summary:
When increasing the value displayed in a spinbox, normally the least
significant digit should stay in place to avoid unecessary movement.
This means the numbers should be right-aligned, instead of being
left-aligned like they are currently.

Since right-aligning still looks a bit unbalanced, centering should be a
good middle ground. In addition, when the number of digits does not
change it makes more sense to pad the number on both sides equally in
any case.

Test Plan:
Check that the spinboxes in the {nav Crop} and {nav Reduce Red Eye}
tools look more pleasant while resizing the crop rectangle and moving
the size slider respectively. For a more profound effect, use a large
image.

Reviewers: #gwenview, muhlenpfordt

Reviewed By: #gwenview, muhlenpfordt

Subscribers: muhlenpfordt

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D14954