Fix Crop tool Aspect ratio sometimes randomly preselected
Summary:
If {nav Crop > Advanced settings > Aspect ratio} is not set to any of
the predefined values the `CropRatioIndex` config value is not
initialized when starting {nav Crop}. This sometimes leads to a
wrong preselection of a random ratio.
This patch initializes the internal index variable.

FIXED-IN: 18.08.1

Test Plan:
* Open {nav Crop} with empty {nav Aspect ratio} setting
* Restart Gwenview / {nav Crop} multiple times
* Check that {nav Aspect ratio} is not randomly preselected
* Check that a selected {nav Aspect ratio} is (re-)stored correctly
* Check that `gwenviewrc` does not store any random `CropRatioIndex`

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D14969