Fix Enter key not working for new Close button in Red Eye Reduce tool
Summary:
When df8cb1abf5ba was merged from stable to master, duplicating the
`QDialogButtonBox` in the {nav Reduce Red Eye} toolbar as done in
0c678376131d was not taken into account, so pressing {key Enter} when
the {nav Close} button had focus for the initial `helpDialogButtonBox`
had no effect anymore.

This can be fixed by always operating on the buttons that are visible.

Test Plan:
Open the {nav Reduce Red Eye} tool and press {key Enter} on the
{nav Close}, {nav Reduce Red Eye} and {nav Cancel} buttons as well as when
none of those buttons is focused.

Reviewers: muhlenpfordt

Reviewed By: muhlenpfordt

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D15014