Use FlowLayout for Crop toolbar
Summary:
The Crop toolbar with enabled {nav Advanced settings} is very wide
and it's not possible to make the Gwenview window smaller with this
toolbar visible.
Using a FlowLayout for the Crop toolbar avoids this issue and breaks
the bar into multiple lines if there is not enough horizontal space.

Before:
{F6251679}

After:
{F6251681}

Test Plan:
* Check Crop toolbar behaviour while resizing Gwenview window
* Check Crop toolbar with {nav Advanced settings} en-/disabled
* Check RedEye toolbar (uses the same `SlideContainer`)
* Check SaveBar after modifying images (uses `SlideContainer`)
* Check Browse Mode {nav Add Filter} (uses `FlowLayout`)

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: shubham, loh.tar, ngraham

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D15398