Ensure the loading indicator is actually displayed
Summary:
Previously the indicator was only displayed if the file's MIME type has
not been recognized yet. However for local files the MIME type was always
detected during a synchronous load by the UI thread, thus the indicator
was never displayed. This can be especially noticed when loading big
panorama images that take a few seconds before fully decoded. With this
change the indicator is displayed as soon as a new url is opened.

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: ngraham

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D21163