Remove redundant QImage::loadFromData()
Summary: The only purpose of this call is to check whether an embedded preview image of a RAW file is smaller than an (arbitrarily chosen) threshold. If this were true, it would be pointless to ask for an image that is only half as big as the one already retrieved.

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: ngraham, #gwenview

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D21285