Accelerate thumnailgenerator when loading RAW files
Summary:
This removes a useless call to QImage::loadFromData() that speeds up the thumbnail generation for RAW files. Its only purpose was to look up the size of the embedded preview image which is useless because

* RAWs have an embedded thumbnail which will be preferred to the embedded preview image (provided the thumb is available and big enough), and
* if the embedded thumbnail cannot be used, QImageReader::read() will decode the embedded preview image again, check its size, scale it, etc..

Also removes a redundant comparison against `MIN_PREV_SIZE` as this is already assured with `pixelSize`.

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: ngraham, #gwenview

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D21284