After import, change default action to "keep" instead of "delete".
Summary:
After importing photos, a dialog pops up which asks whether to
delete or keep the photos on the device/folder from which they
were imported.  Currently, "delete" is the default action.
However, this is quite dangerous for users that don't pay
attention.  Better to present "keep" as the default action.

FEATURE: 337651
FIXED-IN: 19.12.0

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: ngraham

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D24881